+++ src/http-message/src/Helper/HttpHelper.php
@@ -2,7 +2,94 @@
 
 namespace Swoft\Http\Server\Helper;
 
+use Psr\Http\Message\UploadedFileInterface;
+use Swoft\Http\Message\Upload\UploadedFile;
+
+/**
+ * Class HttpHelper
+ */
 class HttpHelper
 {
+    /**
+     * Return an UploadedFile instance array.
+     *
+     * @param array $files A array which respect $_FILES structure
+     *
+     * @throws \InvalidArgumentException for unrecognized values
+     * @return array
+     */
+    public static function normalizeFiles(array $files): array
+    {
+        $normalized = [];
 
+        foreach ($files as $key => $value) {
+            if ($value instanceof UploadedFileInterface) {
+                $normalized[$key] = $value;
+            } elseif (\is_array($value)) {
+                if (isset($value['tmp_name'])) {
+                    $normalized[$key] = self::createUploadedFileFromSpec($value);
+                    continue;
+                }
+
+                $normalized[$key] = self::normalizeFiles($value);
+            } else {
+                throw new \InvalidArgumentException('Invalid value in files specification');
+            }
+        }
+
+        return $normalized;
+    }
+
+
+    /**
+     * Create and return an UploadedFile instance from a $_FILES specification.
+     * If the specification represents an array of values, this method will
+     * delegate to normalizeNestedFileSpec() and return that return value.
+     *
+     * @param array $value $_FILES structure
+     *
+     * @return array|UploadedFileInterface
+     */
+    private static function createUploadedFileFromSpec(array $value)
+    {
+        if (\is_array($value['tmp_name'])) {
+            return self::normalizeNestedFileSpec($value);
+        }
+
+        return new UploadedFile(
+            $value['tmp_name'],
+            (int)$value['size'],
+            (int)$value['error'],
+            $value['name'],
+            $value['type']
+        );
+    }
+
+    /**
+     * Normalize an array of file specifications.
+     * Loops through all nested files and returns a normalized array of
+     * UploadedFileInterface instances.
+     *
+     * @param array $files
+     *
+     * @return UploadedFileInterface[]
+     */
+    private static function normalizeNestedFileSpec(array $files = []): array
+    {
+        $normalizedFiles = [];
+
+        foreach (\array_keys($files['tmp_name']) as $key) {
+            $spec = [
+                'tmp_name' => $files['tmp_name'][$key],
+                'size'     => $files['size'][$key],
+                'error'    => $files['error'][$key],
+                'name'     => $files['name'][$key],
+                'type'     => $files['type'][$key],
+            ];
+
+            $normalizedFiles[$key] = self::createUploadedFileFromSpec($spec);
+        }
+
+        return $normalizedFiles;
+    }
 }