+++ src/http-server/src/Annotation/Parser/RequestMappingParser.php
@@ -6,7 +6,7 @@ use Swoft\Annotation\Annotation\Mapping\AnnotationParser;
 use Swoft\Annotation\Annotation\Parser\Parser;
 use Swoft\Annotation\AnnotationException;
 use Swoft\Http\Server\Annotation\Mapping\RequestMapping;
-use Swoft\Http\Server\Router\RouterRegister;
+use Swoft\Http\Server\Router\RouteRegister;
 
 /**
  * Class RequestMappingParser
@@ -38,7 +38,7 @@ class RequestMappingParser extends Parser
         ];
 
         // Add route info for controller action
-        RouterRegister::addRoute($this->className, $routeInfo);
+        RouteRegister::addRoute($this->className, $routeInfo);
 
         return [];
     }