+++ src/test/src/Benchmark.php
@@ -2,6 +2,10 @@
 
 namespace Swoft\Test;
 
+/**
+ * Class Benchmark
+ * @since 2.0
+ */
 class Benchmark
 {